Logic: Refactor Hint Parsing into Hint Parser
