+++ src/main/java/seedu/address/logic/parser/FindCommandParser.java
@@ -1,15 +1,14 @@
 package seedu.address.logic.parser;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
+import static seedu.address.logic.parser.CliSyntax.LIST_OF_PREFIXES;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_ADDRESS;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_EMAIL;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_EMPTY;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_PHONE;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_TAG;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.function.Predicate;
 
@@ -26,12 +25,6 @@ import seedu.address.model.person.TagsContainKeywordPredicate;
  * Parses input arguments and creates a new FindCommand object
  */
 public class FindCommandParser implements Parser<FindCommand> {
-
-
-
-    private static List<Prefix> prefixList =
-            Arrays.asList(PREFIX_NAME, PREFIX_PHONE, PREFIX_EMAIL, PREFIX_ADDRESS, PREFIX_TAG, PREFIX_EMPTY);
-
     /**
      * Parses the given {@code String} of arguments in the context of the FindCommand
      * and returns an FindCommand object for execution.
@@ -62,7 +55,8 @@ public class FindCommandParser implements Parser<FindCommand> {
                         PREFIX_NAME, PREFIX_PHONE, PREFIX_EMAIL, PREFIX_ADDRESS, PREFIX_TAG);
 
         List<Predicate> predicateList = new ArrayList<>();
-        for (Prefix prefix : prefixList) {
+
+        for (Prefix prefix : LIST_OF_PREFIXES) {
             for (String value : argumentMultimap.getAllValues(prefix)) {
                 if (!value.equals("")) {
                     predicateList.add(valueAndPrefixIntoPredicate(value.trim(), prefix));
